+++ src/main/java/controller/FaviconController.java
@@ -2,6 +2,7 @@ package controller;
 
 import annotation.web.RequestMethod;
 import utils.FileIoUtils;
+import webserver.Controller;
 import webserver.http.HttpRequest;
 import webserver.http.HttpResponse;
