+++ src/main/java/webserver/RequestHandler.java
@@ -5,11 +5,9 @@ import java.net.Socket;
 import java.net.URISyntaxException;
 import java.util.Optional;
 
-import controller.Handler;
-import controller.HandlerMapper;
-import http.HttpRequest;
-import http.HttpRequestParser;
-import http.HttpResponse;
+import webserver.http.HttpRequest;
+import webserver.http.HttpRequestParser;
+import webserver.http.HttpResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import utils.IOUtils;
@@ -18,6 +16,7 @@ public class RequestHandler implements Runnable {
     private static final Logger logger = LoggerFactory.getLogger(RequestHandler.class);
 
     private Socket connection;
+    private ControllerMapper controllerMapper = new ControllerMapper();
 
     public RequestHandler(Socket connectionSocket) {
         this.connection = connectionSocket;
@@ -32,9 +31,10 @@ public class RequestHandler implements Runnable {
             HttpRequest httpRequest = HttpRequestParser.getRequest(request);
             DataOutputStream dos = new DataOutputStream(out);
 
-            Optional<Handler> handler = HandlerMapper.findHandler(httpRequest);
-            if (handler.isPresent()) {
-                HttpResponse response = handler.get().handleRequest(httpRequest);
+            Optional<Controller> controller = controllerMapper.findController(httpRequest);
+
+            if (controller.isPresent()) {
+                HttpResponse response = controller.get().handleRequest(httpRequest);
                 response.sendResponse(dos);
             }
         } catch (IOException | URISyntaxException e) {
