+++ src/test/java/http/HttpRequestHeadersTest.java
@@ -2,6 +2,7 @@ package http;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import webserver.http.HttpRequestHeaders;
 
 import static org.assertj.core.api.Assertions.assertThat;
